parser/requirements.txt
beautifulsoup4
requests
Psycopg2-binary
sockets
python-dotenv
